+++ packages/plugin-packages/mashroom-http-proxy/src/proxy/mashroom-bootstrap-services.ts
@@ -13,13 +13,18 @@ import type {MashroomServicesPluginBootstrapFunction} from '@mashroom/mashroom/t
 import type {Proxy} from '../../type-definitions/internal';
 
 const bootstrap: MashroomServicesPluginBootstrapFunction = async (pluginName, pluginConfig, pluginContextHolder) => {
-    const {proxyImpl, forwardMethods = [], forwardHeaders = [], rejectUnauthorized, poolMaxSockets, socketTimeoutMs, keepAlive, retryOnReset} = pluginConfig;
+    const {
+        proxyImpl, forwardMethods = [], forwardHeaders = [], rejectUnauthorized,
+        /* deprecated */ poolMaxSockets, poolMaxTotalSockets, poolMaxSocketsPerHost, poolMaxWaitingRequestsPerHost,
+        socketTimeoutMs, keepAlive, retryOnReset
+    } = pluginConfig;
     const pluginContext = pluginContextHolder.getPluginContext();
     const logger = pluginContext.loggerFactory('mashroom.httpProxy');
 
     setPoolConfig({
         keepAlive,
-        maxSockets: poolMaxSockets,
+        maxTotalSockets: poolMaxTotalSockets,
+        maxSocketsPerHost: poolMaxSocketsPerHost || /* deprecated */ poolMaxSockets,
         rejectUnauthorized,
     });
 
@@ -33,7 +38,7 @@ const bootstrap: MashroomServicesPluginBootstrapFunction = async (pluginName, pl
         logger.info('Using http-proxy impl based on "node-http-proxy"');
         proxy = new ProxyImplNodeHttpProxy(socketTimeoutMs, rejectUnauthorized, interceptorHandler, headerFilter, retryOnReset, pluginContext.loggerFactory);
     }
-    const service = new MashroomHttpProxyService(forwardMethods, proxy);
+    const service = new MashroomHttpProxyService(forwardMethods, proxy, poolMaxWaitingRequestsPerHost);
 
     startExportPoolMetrics(pluginContextHolder);
     startExportRequestMetrics(proxy, pluginContextHolder);