+++ packages/plugin-packages/mashroom-http-proxy/type-definitions/internal.ts
@@ -13,7 +13,8 @@ export interface HttpHeaderFilter {
 export type PoolConfig = {
     keepAlive: boolean;
     rejectUnauthorized: boolean;
-    maxSockets: number;
+    maxSocketsPerHost: number;
+    maxTotalSockets: number | null;
 }
 
 export type PoolMetrics = {